requirements.txt
asgiref==3.2.7
Django==2.2.10
django-activity-stream==0.8.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
